implements start auction usecase

+++ src/auctionservice/pkg/models/auction.go
@@ -8,5 +8,5 @@ type Auction struct {
 //AuctionResponse is the model of the reponse from the auction
 type AuctionResponse struct {
 	BidderID    string  `json:"bidder_id"`
-	MaxBidValue float32 `json:"price"`
+	MaxBidValue float64 `json:"price"`
 }
